backend/requirements.txt
fastapi 
uvicorn
pydantic
openai
pandas 
python-dotenv
typing 
datetime
yfinance
requests 
redis
loguru
python-multipart